inventory_management/requirements.txt
Django>=3.0
djangorestframework>=3.12
djangorestframework-simplejwt>=4.6
psycopg2-binary>=2.9
redis>=4.0
django-redis>=4.12

